Enhance training_helper_MixTS.py with Annoy integration for similarity graph construction, optimize clean_selection function, and update README with dataset download instructions. Modify .gitignore and adjust data type handling in load_ucr_pre.py.

+++ src/ucr_data/load_ucr_pre.py
@@ -49,8 +49,8 @@ def to_time_series(ts, remove_nans=False):
     ts_out = numpy.array(ts, copy=True)
     if ts_out.ndim <= 1:
         ts_out = ts_out.reshape((-1, 1))
-    if ts_out.dtype != numpy.float:
-        ts_out = ts_out.astype(numpy.float)
+    if ts_out.dtype != float:
+        ts_out = ts_out.astype(float)
     if remove_nans:
         ts_out = ts_out[:ts_size(ts_out)]
     return ts_out
@@ -97,7 +97,7 @@ def ts_size(ts):
     return sz
 
 
-def to_time_series_dataset(dataset, dtype=numpy.float):
+def to_time_series_dataset(dataset, dtype=float):
     """Transforms a time series dataset so that it fits the format used in
     ``tslearn`` models.
 
@@ -106,7 +106,7 @@ def to_time_series_dataset(dataset, dtype=numpy.float):
     dataset : array-like
         The dataset of time series to be transformed. A single time series will
         be automatically wrapped into a dataset with a single entry.
-    dtype : data type (default: numpy.float)
+    dtype : data type (default: float)
         Data type for the returned dataset.
 
     Returns
@@ -182,7 +182,7 @@ def load_txt_uea(dataset_path):
     """
     data = numpy.loadtxt(dataset_path)
     X = to_time_series_dataset(data[:, 1:])
-    y = data[:, 0].astype(numpy.int)
+    y = data[:, 0].astype(int)
     return X, y
